Rename app.py file name because of conflict
